% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graysacle-package.R
\docType{package}
\name{graysacle-package}
\alias{graysacle}
\alias{graysacle-package}
\title{graysacle: Graysacle theme}
\description{
Allow us to use the graysacle theme inside a ShinyApp.
}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan@thinkr.fr}

}
\keyword{internal}
