% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graysacle_card_txt}
\alias{graysacle_card_txt}
\title{Graysacle card with text}
\usage{
graysacle_card_txt(title, description)
}
\arguments{
\item{title}{title}

\item{description}{description}
}
\value{
tagList
}
\description{
Graysacle card with text
}
