% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{graysacle_header}
\alias{graysacle_header}
\title{Header}
\usage{
graysacle_header(
  ref_to_go,
  title = "Grayscale",
  description = "Example of grayscale",
  img = "www/assets/img/bg-masthead.jpg"
)
}
\arguments{
\item{ref_to_go}{ref of section to go when click on Get Started}

\item{title}{title}

\item{description}{description}

\item{img}{path to a img (storage in www folder)}
}
\value{
tagList
}
\description{
Header
}
