% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graysacle_one_project}
\alias{graysacle_one_project}
\title{One project div}
\usage{
graysacle_one_project(src_to_img, title, description, dark = FALSE)
}
\arguments{
\item{src_to_img}{path to www folder for img}

\item{title}{title}

\item{description}{description}

\item{dark}{if TRUE black background otherwise white}
}
\value{
tagList
}
\description{
One project div
}
