% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{graysacle_page}
\alias{graysacle_page}
\title{Get graysacle page}
\usage{
graysacle_page(header, nav, body, footer)
}
\arguments{
\item{header}{result of graysacle_header()}

\item{nav}{result of graysacle_nav()}

\item{body}{html code}

\item{footer}{html code}
}
\value{
tagList
}
\description{
Get graysacle page
}
