% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graysacle_projects_img}
\alias{graysacle_projects_img}
\title{Another formatting project with image}
\usage{
graysacle_projects_img(img, title, description, img_first = TRUE)
}
\arguments{
\item{img}{path to www folder for img}

\item{title}{title}

\item{description}{description}

\item{img_first}{if TRUE, img should be on the left side}
}
\value{
tagList
}
\description{
Better using this fonction if you have more then one project
}
