% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graysacle_section}
\alias{graysacle_section}
\title{Section}
\usage{
graysacle_section(id, ...)
}
\arguments{
\item{id}{one of these id pass to graysacle_nav}

\item{...}{html code}
}
\value{
tagList
}
\description{
Section
}
