% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graysacle_section_bg}
\alias{graysacle_section_bg}
\title{Section and choose background dark or not}
\usage{
graysacle_section_bg(id, ..., dark = FALSE)
}
\arguments{
\item{id}{one of these id pass to graysacle_nav}

\item{...}{html code}

\item{dark}{if TRUE black background otherwise white}
}
\value{
tagList
}
\description{
Section and choose background dark or not
}
