% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graysacle_projects}
\alias{graysacle_projects}
\title{Another formatting project}
\usage{
graysacle_projects(..., title, description, first = TRUE)
}
\arguments{
\item{...}{html code (like ggplot or table)}

\item{title}{title}

\item{description}{description}

\item{first}{if TRUE, your html code should be on the left side}
}
\value{
tagList
}
\description{
Better using this fonction if you have more then one project
}
